# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Laravel API Documentation",
        "description": "Swagger documentation for my Laravel project",
        "version": "1.0.0"
    },
    "paths": {
        "/api/podcasts/{podcast_id}/episodes": {
            "get": {
                "tags": [
                    "Episodes"
                ],
                "summary": "Get all episodes of a podcast",
                "operationId": "c69fe57d6278ec456b66f447e361abc5",
                "parameters": [
                    {
                        "name": "podcast_id",
                        "in": "path",
                        "description": "Podcast ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of episodes"
                    },
                    "404": {
                        "description": "Podcast not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Episodes"
                ],
                "summary": "Create a new episode",
                "operationId": "b392de7d6ef50873e1819ac47e497ef3",
                "parameters": [
                    {
                        "name": "podcast_id",
                        "in": "path",
                        "description": "Podcast ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "audio"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Episode 1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Episode description"
                                    },
                                    "audio": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Episode created successfully"
                    }
                }
            }
        },
        "/api/episodes/{id}": {
            "get": {
                "tags": [
                    "Episodes"
                ],
                "summary": "Get a specific episode",
                "operationId": "682c9f0b582614c5d25814ad0a6659c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Episode ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Episode details"
                    },
                    "404": {
                        "description": "Episode not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Episodes"
                ],
                "summary": "Update an episode",
                "operationId": "7c242aee03c0c562d61e184cc40ffbcf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Episode ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Episode"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    },
                                    "audio": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Episode updated successfully"
                    },
                    "404": {
                        "description": "Episode not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Episodes"
                ],
                "summary": "Delete an episode",
                "operationId": "2b31a0a0e3d3c3f435a21fe63779b608",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Episode ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Episode deleted successfully"
                    },
                    "404": {
                        "description": "Episode not found"
                    }
                }
            }
        },
...
    },
...
}